requirements.txt
# TODO fix the version for all packages.
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.88.0
google-genai
pydantic>=2.0.0
openai==1.3.0
requests==2.31.0
python-dotenv
twilio
msal
pyyaml